requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
argh==0.26.2              # via watchdog
click==6.7                # via python-dotenv
pathtools==0.1.2          # via watchdog
python-dotenv==0.7.1
pyyaml==3.12              # via watchdog
twitter==1.18.0
watchdog==0.8.3
